+++ services/workflows-service/src/document/utils/fetch-csv-from-url-and-convert-to-base-64.ts
@@ -1,0 +1,14 @@
+import axios from 'axios';
+
+export const fetchCsvFromUrlAndConvertToBase64 = async (csvUrl: string) => {
+  const response = await axios.get(csvUrl, {
+    responseType: 'arraybuffer',
+  });
+  const buffer = response.data;
+  const base64 = Buffer.from(buffer).toString('base64');
+  const contentType = response.headers['content-type'];
+
+  const base64Result = `data:${contentType};base64,${base64}`;
+
+  return base64Result;
+};
