+++ services/workflows-service/src/document/utils/is-csv-mimetype.ts
@@ -1,0 +1,2 @@
+export const isCsvMimetype = (mimeType: string | null) =>
+  mimeType === 'text/csv' || mimeType === 'application/csv';
